# Optimize `string-append`