# Hello world benchmark

```
> hyperfine -w 10 '~/worktree/7a1181edfad9f3e5/target/release/stak ~/foo.scm' 'target/release/stak ~/foo.scm'
Benchmark 1: ~/worktree/7a1181edfad9f3e5/target/release/stak ~/foo.scm
  Time (mean ± σ):      1.805 s ±  0.003 s    [User: 1.792 s, System: 0.004 s]
  Range (min … max):    1.802 s …  1.810 s    10 runs

Benchmark 2: target/release/stak ~/foo.scm
  Time (mean ± σ):      1.787 s ±  0.004 s    [User: 1.775 s, System: 0.004 s]
  Range (min … max):    1.783 s …  1.797 s    10 runs

Summary
  target/release/stak ~/foo.scm ran
    1.01 ± 0.00 times faster than ~/worktree/7a1181edfad9f3e5/target/release/stak ~/foo.scm
```

This is not that fast...

# GC benchmark

~~A number of GC is 4 times smaller than before in the self-hosted compiler.~~ I didn't rebase to the `main` branch.

## Before

```sh
> cargo run --release --bin stak compile.scm < prelude.scm < compile.scm > out.bc 2> gc.log && wc gc.log
   12112   12134   36824 gc.log
```

## After

```sh
> cargo run --release --bin stak compile.scm < prelude.scm < compile.scm > out.bc 2> gc.log && wc gc.log
   11707   11729   35609 gc.log
```

Even after a0068f4c654be99abb9ae826dad924c57813cf2c,

```
> cargo run --release --bin stak compile.scm < prelude.scm < compile.scm > out.bc 2> gc.log && wc gc.log
   11594   11610   35112 gc.log
```